virtio-gpu: Add an option to connect all outputs on startup
When multiple outputs are enabled using the "max_outputs" attribute,
only the first connector appears as "Connected" in the guest DRM
device.  Additional connectors must be enabled from the host side
UI frontend before they are usable by the guest.  However, multiple
outputs can still be of use on a headless configuration, if for example,
the display will only be used for taking periodic screenshots in the
guest or integration into a CI pipeline, etc.

Add an option to start all of the outputs in the "Connected" state,
so that they are immediately available to the guest.

Resolves: https://gitlab.com/qemu-project/qemu/-/issues/1107
Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>